+++ tests/_bootstrap.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * PHP-CS-Fixer plugin for HiDev
+ *
+ * @link      https://github.com/hiqdev/hidev-php-cs-fixer
+ * @package   hidev-php-cs-fixer
+ * @license   BSD-3-Clause
+ * @copyright Copyright (c) 2015, HiQDev (http://hiqdev.com/)
+ */
+
+error_reporting(-1);
+
+define('YII_ENABLE_ERROR_HANDLER', false);
+define('YII_DEBUG', true);
+$_SERVER['SCRIPT_NAME']     = '/' . __DIR__;
+$_SERVER['SCRIPT_FILENAME'] = __FILE__;
+require_once __DIR__ . '/../vendor/autoload.php';
+require_once __DIR__ . '/../vendor/yiisoft/yii2/Yii.php';