+++ CHANGELOG.md
@@ -3,7 +3,7 @@ hiqdev/hidev-php-cs-fixer changelog
 
 ## Under development
 
-- Added Travis CI
+- Added tests and Travis CI
 
 ## 0.1.0 2015-11-09